requirements.txt
streamlit==1.16.0
datasets==2.6.0
pandas==1.5.1
numpy==1.23.3
matplotlib==3.7.0
plotly==5.13.0
Pillow